requirements.txt
streamlit==1.32.0
pymongo==4.6.2
python-dotenv==1.0.1
bcrypt==4.1.2 